# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vector Portal API
  version: 1.0.0
  description: |
    Document-workflow portal for multi-role approval chains.
    Manages events (document packages), their sections, and a
    role-based submit/approve/return pipeline with file attachments,
    inline comments, and a completed-document library.

    **Roles** (highest to lowest): ADMIN, PROTOCOL, DEPUTY, SUPERVISOR,
    SUPER_COLLABORATOR, COLLABORATOR, ANALYST.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by POST /api/auth/login (24 h expiry)
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]

    # ── Auth ──────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        isExternal:
          type: boolean
        entityName:
          type: string
          nullable: true
        mustChangePassword:
          type: boolean
    UserRole:
      type: string
      enum: [ADMIN, PROTOCOL, DEPUTY, SUPERVISOR, SUPER_COLLABORATOR, COLLABORATOR, ANALYST]
...
    Country:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        code:
          type: string
...
paths:
...
  /api/users:
    get:
      tags: [users]
      summary: List all users (ADMIN only)
      responses:
        '200':
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
    post:
      tags: [users]
      summary: Create a user (ADMIN only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, username, email, password, role]
              properties:
                fullName:
                  type: string
                  maxLength: 200
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  $ref: '#/components/schemas/UserRole'
                departmentId:
                  type: integer
                  nullable: true
                isExternal:
                  type: boolean
                entityName:
                  type: string
                  nullable: true
                countryIds:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}:
    patch:
      tags: [users]
      summary: Update a user (ADMIN only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  format: email
                role:
                  $ref: '#/components/schemas/UserRole'
                departmentId:
                  type: integer
                  nullable: true
                isExternal:
                  type: boolean
                entityName:
                  type: string
                password:
                  type: string
                  minLength: 8
                countryIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: User not found
        '409':
          description: Username conflict
  /api/users/{id}/countries:
    get:
      tags: [users]
      summary: Get country assignments for a user (ADMIN only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Array of assigned countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'

  # ── Departments ───────────────────────────────────────────────────
...